--- build.gradle
@@ -1,94 +1,0 @@
-import org.apache.tools.ant.filters.ReplaceTokens
-
-plugins {
-    id 'com.github.johnrengelman.shadow' version '8.1.1'
-    id 'eclipse'
-    id 'java'
-}
-
-compileJava.options.encoding = 'Cp1252'
-compileTestJava.options.encoding = 'Cp1252'
-
-compileJava {
-    sourceCompatibility = '1.17'
-    targetCompatibility = '1.17'
-}
-
-group = 'me.Asleepp'
-version = '1.5'
-
-repositories {
-    mavenCentral()
-    maven {
-        name = "papermc-repo"
-        url = "https://repo.papermc.io/repository/maven-public/"
-    }
-    maven {
-        name = "sonatype"
-        url = "https://oss.sonatype.org/content/groups/public/"
-    }
-    // Skript
-    maven {
-        name = "Skript"
-        url = 'https://repo.skriptlang.org/releases'
-    }
-    // ItemsAdder
-    maven {
-        name = "ItemsAdder"
-        url = 'https://jitpack.io'
-
-    }
-}
-
-dependencies {
-    compileOnly "io.papermc.paper:paper-api:1.20.4-R0.1-SNAPSHOT"
-    compileOnly (group: 'com.github.SkriptLang', name: 'Skript', version: '2.7.1') {
-        transitive = false
-    }
-    compileOnly 'com.github.LoneDev6:API-ItemsAdder:3.6.1'
-
-}
-
-def targetJavaVersion = 17
-java {
-    def javaVersion = JavaVersion.toVersion(targetJavaVersion)
-    sourceCompatibility = javaVersion
-    targetCompatibility = javaVersion
-    if (JavaVersion.current() < javaVersion) {
-        toolchain.languageVersion = JavaLanguageVersion.of(targetJavaVersion)
-    }
-}
-
-tasks.withType(JavaCompile).configureEach {
-    options.encoding = 'UTF-8'
-
-    if (targetJavaVersion >= 10 || JavaVersion.current().isJava10Compatible()) {
-        options.release.set(targetJavaVersion)
-    }
-}
-
-build {
-    dependsOn(shadowJar)
-    tasks.withType(JavaCompile).tap {
-        configureEach {
-            options.compilerArgs.add("-Xlint:unchecked")
-            options.compilerArgs.add("-Xlint:deprecation")
-        }
-    }
-}
-
-processResources {
-    filter ReplaceTokens, tokens: ["version": project.version]
-    }
-
-
-shadowJar {
-    dependencies {
-        include(dependency('org.bstats:bstats-bukkit'))
-        include(dependency('org.bstats:bstats-base'))
-    }
-    relocate 'org.bstats', 'me.asleepp.SkriptItemsAdder.bstats'
-    configurations = [project.configurations.shadow]
-    archiveVersion = project.property("version")
-    minimize()
-}
