+++ build.gradle.kts
@@ -1,0 +1,110 @@
+import org.apache.tools.ant.filters.ReplaceTokens
+
+plugins {
+    `java-library`
+    `maven-publish`
+    alias(libs.plugins.shadow)
+    alias(libs.plugins.run.paper)
+}
+
+repositories {
+    mavenLocal()
+    mavenCentral()
+    maven("https://repo.papermc.io/repository/maven-public/")
+    maven("https://repo.skriptlang.org/releases")
+    maven("https://jitpack.io")
+}
+
+java {
+    disableAutoTargetJvm()
+    toolchain.languageVersion = JavaLanguageVersion.of(21)
+}
+
+description = "A skript addon that aims to bring more to linking ItemsAdder and Skript"
+group = "me.Asleepp"
+version = "1.6"
+
+dependencies {
+    compileOnly(libs.paper.api)
+    compileOnly(libs.lombok)
+    compileOnly(libs.itemsadder)
+    compileOnly(libs.skript)
+    annotationProcessor(libs.lombok)
+}
+
+
+tasks {
+    jar {
+        enabled = false
+    }
+
+    shadowJar {
+        archiveFileName = "${rootProject.name}-${project.version}.jar"
+        archiveClassifier = null
+
+        manifest {
+            attributes["Implementation-Version"] = rootProject.version
+        }
+
+        relocate("org.bstats", "me.asleepp.SkriptItemsAdder.bstats")
+        configurations = listOf(project.configurations.shadow.get())
+        minimize()
+    }
+
+    assemble {
+        dependsOn(shadowJar)
+    }
+
+    withType<JavaCompile> {
+        options.encoding = Charsets.UTF_8.name()
+        options.release = 17
+    }
+
+    withType<Javadoc>() {
+        options.encoding = Charsets.UTF_8.name()
+    }
+
+    processResources {
+        filter<ReplaceTokens>("tokens" to mapOf("version" to project.version))
+        inputs.property("version", project.version)
+
+        filesMatching("plugin.yml") {
+            expand(
+                "version" to rootProject.version,
+            )
+        }
+    }
+
+    defaultTasks("build")
+
+    // 1.8.8 - 1.16.5 = Java 8
+    // 1.17           = Java 16
+    // 1.18 - 1.20.4  = Java 17
+    // 1-20.5+        = Java 21
+    val version = "1.20.4"
+    val javaVersion = JavaLanguageVersion.of(17)
+
+    val jvmArgsExternal = listOf(
+        "-Dcom.mojang.eula.agree=true"
+    )
+
+    runServer {
+        minecraftVersion(version)
+        runDirectory = rootDir.resolve("run/paper/$version")
+
+        javaLauncher = project.javaToolchains.launcherFor {
+            languageVersion = javaVersion
+        }
+
+        downloadPlugins {
+            url("https://download.luckperms.net/1552/bukkit/loader/LuckPerms-Bukkit-5.4.137.jar")
+            url("https://ci.lucko.me/job/spark/439/artifact/spark-bukkit/build/libs/spark-1.10.93-bukkit.jar")
+            url("https://github.com/SkriptLang/Skript/releases/download/2.9.1/Skript-2.9.1.jar")
+            url("https://github.com/dmulloy2/ProtocolLib/releases/download/5.2.0/ProtocolLib.jar")
+            url("https://github.com/LoneDev6/SpigotUtilities/raw/master/LoneLibs.jar")
+        }
+
+        jvmArgs = jvmArgsExternal
+    }
+}
+
