+++ src/main/java/me/asleepp/SkriptItemsAdder/elements/effects/blocks/EffReplaceBlock.java
@@ -1,4 +1,4 @@
-package me.asleepp.SkriptItemsAdder.elements.effects;
+package me.asleepp.SkriptItemsAdder.elements.effects.blocks;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.doc.Description;
