+++ src/main/java/me/asleepp/SkriptItemsAdder/other/aliases/AliasesGenerator.java
@@ -1,4 +1,4 @@
-package me.asleepp.SkriptItemsAdder.other;
+package me.asleepp.SkriptItemsAdder.other.aliases;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.aliases.Aliases;
@@ -13,20 +13,23 @@ import org.bukkit.configuration.file.YamlConfiguration;
 import java.io.File;
 import java.io.IOException;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 
 public class AliasesGenerator {
 
     private final SkriptItemsAdder plugin;
     private final File aliasesFile;
     private FileConfiguration aliasesConfig;
-    private final Map<String, String> itemAliases = new HashMap<>();
-    private final Set<String> generatedAliases = new HashSet<>();
+    private final Map<String, String> itemAliases = new ConcurrentHashMap<>();
+    private final Set<String> generatedAliases = ConcurrentHashMap.newKeySet();
 
     public AliasesGenerator(SkriptItemsAdder plugin) {
         this.plugin = plugin;
         this.aliasesFile = new File(plugin.getDataFolder(), "aliases.yml");
         loadAliasesFile();
         loadAliasesFromFile();
+        //watchAliasesFile();
     }
 
     public void loadAliasesFile() {
@@ -37,7 +40,7 @@ public class AliasesGenerator {
         aliasesConfig = YamlConfiguration.loadConfiguration(aliasesFile);
     }
 
-    private void loadAliasesFromFile() {
+    public void loadAliasesFromFile() {
         if (aliasesConfig.contains("items")) {
             for (String key : aliasesConfig.getConfigurationSection("items").getKeys(false)) {
                 itemAliases.put(key, aliasesConfig.getString("items." + key));
@@ -86,35 +89,70 @@ public class AliasesGenerator {
         try {
             addonProvider.addAlias(aliasName, namespacedId, null, new HashMap<>());
             plugin.getLogger().info("Registered alias: " + alias + " for item: " + namespacedId + " with AliasesProvider.");
-        } catch (InvalidMinecraftIdException e) {
-            plugin.getLogger().severe("Invalid Minecraft ID: " + namespacedId);
-        } catch (NullPointerException e) {
-            plugin.getLogger().severe("Failed to register alias with AliasesProvider: " + e.getMessage());
+        } catch (InvalidMinecraftIdException ignored) {
         }
-    }
-
-    public void generateAliasesForNewItems(List<CustomStack> newItems) {
-        for (CustomStack item : newItems) {
-            String alias = generateAliasForItem(item);
-            registerAlias(alias, item.getNamespacedID());
-        }
-        saveAliasesToFile();
     }
 
     public void generateAliasesForAllItems() {
         plugin.getLogger().info("Generating aliases for all items.");
         List<CustomStack> allItems = ItemsAdder.getAllItems();
         plugin.getLogger().info("Total items: " + allItems.size());
+
+        Set<String> currentNamespaces = allItems.stream()
+                .map(CustomStack::getNamespacedID)
+                .collect(Collectors.toSet());
+
+        // Check if aliases need to be generated or updated
+        boolean allAliasesExist = currentNamespaces.stream()
+                .allMatch(itemAliases::containsValue);
+
+        if (allAliasesExist && currentNamespaces.size() == itemAliases.size()) {
+            plugin.getLogger().info("All aliases are up to date.");
+            return;
+        }
+
         for (CustomStack item : allItems) {
-            if (!itemAliases.containsValue(item.getNamespacedID())) {
+            String namespacedId = item.getNamespacedID();
+            if (!itemAliases.containsValue(namespacedId)) {
                 String alias = generateAliasForItem(item);
-                plugin.getLogger().info("Generated alias: " + alias + " for item: " + item.getNamespacedID());
-                registerAlias(alias, item.getNamespacedID());
+                registerAlias(alias, namespacedId);
             }
         }
-        saveAliasesToFile();
-        plugin.getLogger().info("Aliases have been generated");
+
+        Set<String> unusedAliases = itemAliases.keySet().stream()
+                .filter(alias -> !generatedAliases.contains(alias))
+                .collect(Collectors.toSet());
+        for (String unusedAlias : unusedAliases) {
+            itemAliases.remove(unusedAlias);
+        }
     }
+
+//    private void watchAliasesFile() {
+//        try {
+//            WatchService watchService = FileSystems.getDefault().newWatchService();
+//            Path path = aliasesFile.toPath().getParent();
+//            path.register(watchService, StandardWatchEventKinds.ENTRY_MODIFY);
+//
+//            Bukkit.getScheduler().runTaskAsynchronously(plugin, () -> {
+//                try {
+//                    while (true) {
+//                        WatchKey key = watchService.take();
+//                        for (WatchEvent<?> event : key.pollEvents()) {
+//                            if (event.context().toString().equals(aliasesFile.getName())) {
+//                                Bukkit.getScheduler().runTask(plugin, this::loadAliasesFromFile);
+//                                syncAliasesWithProvider();
+//                            }
+//                        }
+//                        key.reset();
+//                    }
+//                } catch (InterruptedException ignored) {
+//                }
+//            });
+//        } catch (IOException e) {
+//            plugin.getLogger().severe("Failed to watch aliases.yml for changes: " + e.getMessage());
+//            e.printStackTrace();
+//        }
+//    }
 
     public void syncAliasesWithProvider() {
         AliasesProvider addonProvider = Aliases.getAddonProvider(Skript.getAddonInstance());
@@ -124,46 +162,14 @@ public class AliasesGenerator {
             AliasesProvider.AliasName aliasName = new AliasesProvider.AliasName(entry.getKey(), entry.getKey(), 0);
             try {
                 addonProvider.addAlias(aliasName, entry.getValue(), null, new HashMap<>());
-            } catch (InvalidMinecraftIdException e) {
-                plugin.getLogger().severe("Failed to sync aliases! " + e.getMessage());
+            } catch (InvalidMinecraftIdException ignored) {
             }
         }
     }
 
-    public void handleAliasChange(String alias, String newNamespacedId) {
-        itemAliases.put(alias, newNamespacedId);
-        aliasesConfig.set("items." + alias, newNamespacedId);
-        saveAliasesToFile();
-        syncAliasesWithProvider();
-    }
-
-    public void handleItemUnregister(CustomStack item) {
-        String alias = getAliasForItem(item);
-        if (alias != null) {
-            itemAliases.remove(alias);
-            aliasesConfig.set("items." + alias, null);
-            saveAliasesToFile();
-            syncAliasesWithProvider();
-        }
-    }
-
-    private String getAliasForItem(CustomStack item) {
-        for (Map.Entry<String, String> entry : itemAliases.entrySet()) {
-            if (entry.getValue().equals(item.getNamespacedID())) {
-                return entry.getKey();
-            }
-        }
-        return null;
-    }
-
-    public void handleRemovedItems(List<CustomStack> removedItems) {
-        for (CustomStack item : removedItems) {
-            handleItemUnregister(item);
-        }
-    }
 
     public String getNamespacedId(String alias) {
-        return itemAliases.getOrDefault(alias, alias);
+        return itemAliases.get(alias);
     }
 
     public Set<String> getGeneratedAliases() {
