+++ src/main/java/me/asleepp/SkriptItemsAdder/other/aliases/CustomItemType.java
@@ -1,4 +1,4 @@
-package me.asleepp.SkriptItemsAdder.other;
+package me.asleepp.SkriptItemsAdder.other.aliases;
 
 import org.bukkit.event.Event;
