+++ src/main/resources/plugin.yml
@@ -2,9 +2,10 @@ name: skript-itemsadder
 authors: [Asleepp]
 description: A Skript Addon that adds useful syntax to be used in conjunction with ItemsAdder
 api-version: 1.13
-version: 1.5
+version: ${version}
 main: me.asleepp.SkriptItemsAdder.SkriptItemsAdder
 depend: [Skript, ItemsAdder]
+
 
 commands:
   skriptitemsadder:
